requirements.txt
telethon>=1.28.5
python-dotenv>=1.0.0
pandas>=2.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
Pillow>=10.0.0
python-dateutil>=2.8.2